batch: Extract merge validation
Batch merge still owns presence application, absence suppression, candidate
enumeration, and discard reversal orchestration.

Structural placement validation is a separate gate over line mappings and
claimed runs, so keeping its interval-fact model inside batch.merge hides
the safety boundary from readers.

This commit moves the structural validity checks into
batch.merge_validation and rewires batch.merge to call the new validation
owner.

Tests will follow the new validation module boundary.
